# Conflicts:
#	server/src/main/java/de/ingrid/igeserver/profiles/ingrid/importer/GeodatasetMapper.kt
#	server/src/main/java/de/ingrid/igeserver/profiles/ingrid_krzn/exporter/KrznProfileTransformer.kt

+++ server/src/main/java/de/ingrid/igeserver/profiles/ingrid/importer/GeodatasetMapper.kt
@@ -76,7 +76,7 @@ class GeodatasetMapper(metadata: Metadata, codeListService: CodelistHandler, cat
                     ?: emptyList()
             }
             ?.getOrNull(0)
-            ?.toInt()
+            ?.toDouble()
     }
 
     fun getQualities(): List<Quality> {
@@ -275,7 +275,7 @@ class GeodatasetMapper(metadata: Metadata, codeListService: CodelistHandler, cat
         val nameId = codeListService.getCodeListEntryId(info.codelist, name, "de")
         val nameKeyValue = if (nameId == null) KeyValue(null, name) else KeyValue(nameId)
         val parameter = info.element.measureDescription?.value
-        val value = info.element.result?.dqQuantitativeResult?.value?.getOrNull(0)?.value?.toInt()
+        val value = info.element.result?.dqQuantitativeResult?.value?.getOrNull(0)?.value?.toDouble()
         return Quality(info.type, nameKeyValue, value, parameter)
     }
 
@@ -362,8 +362,8 @@ class GeodatasetMapper(metadata: Metadata, codeListService: CodelistHandler, cat
 
     fun getGeometryContexts(): List<GeometryContextInternal> {
         return metadata.spatialRepresentationInfo
-            ?.mapNotNull {it.mdGeometryContext } 
-            ?.map { 
+            ?.mapNotNull {it.mdGeometryContext }
+            ?.map {
                 val feature = it.geometricFeature?.nominalFeature ?: it.geometricFeature?.ordinalFeature ?: it.geometricFeature?.scalarFeature ?: it.geometricFeature?.otherFeature
                 GeometryContextInternal(
                     feature?.featureName?.value,
@@ -385,7 +385,7 @@ class GeodatasetMapper(metadata: Metadata, codeListService: CodelistHandler, cat
             KeyValue(
                 item?.attributeCode?.value ?: item?.attributeContent?.value,
                 item?.attributeDescription?.value
-                ) 
+            )
         } ?: emptyList()
     }
 
@@ -399,7 +399,7 @@ class GeodatasetMapper(metadata: Metadata, codeListService: CodelistHandler, cat
         }
     }
 
-    private fun getVerticalAbsoluteExternalPositionalAccuracy(): Int? {
+    private fun getVerticalAbsoluteExternalPositionalAccuracy(): Double? {
         return metadata.dataQualityInfo
             ?.flatMap {
                 it.dqDataQuality?.report
@@ -410,10 +410,9 @@ class GeodatasetMapper(metadata: Metadata, codeListService: CodelistHandler, cat
                     } ?: emptyList()
             }
             ?.getOrNull(0)
-            ?.toInt()
+            ?.toDouble()
     }
-
-    private fun getHorizontalAbsoluteExternalPositionalAccuracy(): Int? {
+    private fun getHorizontalAbsoluteExternalPositionalAccuracy(): Double? {
         return metadata.dataQualityInfo
             ?.flatMap {
                 it.dqDataQuality?.report
@@ -424,10 +423,9 @@ class GeodatasetMapper(metadata: Metadata, codeListService: CodelistHandler, cat
                     } ?: emptyList()
             }
             ?.getOrNull(0)
-            ?.toInt()
+            ?.toDouble()
     }
-
-    private fun getGriddedDataPositionalAccuracy(): Int? {
+    private fun getGriddedDataPositionalAccuracy(): Double? {
         return metadata.dataQualityInfo
             ?.flatMap {
                 it.dqDataQuality?.report
@@ -437,7 +435,7 @@ class GeodatasetMapper(metadata: Metadata, codeListService: CodelistHandler, cat
                     } ?: emptyList()
             }
             ?.getOrNull(0)
-            ?.toInt()
+            ?.toDouble()
     }
 }
 
@@ -467,9 +465,9 @@ data class VectorSpatialRepresentation(
 )
 
 data class PositionalAccuracy(
-    val vertical: Int?,
-    val horizontal: Int?,
-    val griddedDataPositionalAccuracy: Int?
+    val vertical: Double?,
+    val horizontal: Double?,
+    val griddedDataPositionalAccuracy: Double?
 )
 
 data class GeometryContextInternal(
